refactor: extract dry-run SQL harness
